Peakcoin Core integration/staging tree
What is Peakcoin?

Peakcoin is an experimental new digital currency that enables instant
payments to anyone, anywhere in the world. Peakcoin uses peer-to-peer
technology to operate with no central authority: managing transactions
and issuing money are carried out collectively by the network. Peakcoin
Core is the name of open source software which enables the use of this
currency.

For more information, as well as an immediately useable, binary version
of the Peakcoin Core software, see https://peakcoin.io